boot_param: fix nits in impl
This CL addresses nits from https://crrev.com/c/6814152:
- makes sure that only a valid BOOT_PARAM_VERSION can be defined
- explicitly defines the size of set_cbor_bstr_hdr16()'s size param

BUG=none
TEST=make build (with various BOOT_PARAM_VERSION defined)

+++ boot_param/boot_param.c
@@ -922,7 +922,7 @@ static inline void set_cbor_bstr_hdr16(
 	/* [OUT] header to be filled */
 	uint8_t hdr[3],
 	/* [IN] size of bstr value */
-	size_t size
+	uint16_t size
 )
 {
 	hdr[0] = CBOR_HDR1(CBOR_MAJOR_BSTR, CBOR_BYTES2);
